Very basic ClipScript support
Clicking "InOut", processes each line of the clipboard and treats the
"Input" text box as a C# snippet that handles one line. Variables
available include line (string content of the current line), number (the
line number) and sb (a StringBuilder you append your text to). Results
are written to the "Output" text box. No saving or anything nice, but it
does work.