ClipScript Editor
Created the basic ClipScript editor. Not working yet, but pretty far
along. This will allow a user to select their source location (input and
destination location (output) as well as a processing script. This
allows combinations like clipboard to clipboard, clipboard to file, file
to file, text box to text box or any combination there of.