Added 'Location' attribute. Fixed urn AllNotIn
I have added a new attribute to all the nodes in the meta data xml
called 'Location'. It will simply be the positional index of the column,
property, table, index, etc. in the list starting at 1. So if there are
3 tables, the first node will have Location=1, second Location=2, etc.
Same with Columns. I found it was useful to have an absolute number
rather than what might be a relative position() in the xsl. The
ClipScript editor is taking shape. No save or load yet, but it runs a
script pretty well. I also fixed a bug in the internal AllNotIn()
function which would break if more than one node had no value.