Command line excel
The quick script exe generated for the command line can now auto export
an excel file to a tab delimited .txt and then process that (or just a
.txt). New command ~~BeginString: variableName and ~~EndString:.
Everything between them is stored in a string variable named
variableName. Intellegent enough to use either @ or a number of
continuous string appends. Implemented as the first just in time GenArea
creation. Enhancements to StringExtensions including
TransformAllNotEmpty which will take a List<string> and perform a lambda
expression across it updating the contents.