~~To: working
What a weird thing this was. Had to manually copy the contents of the now old StreamBuilder to a class that was very close to it yet does not display the omg error Stream == FileStream closes and disposes of the FileStream, every single time... Well not now. Now the new StreamBuilder class works as expected (and as it has been working for years)... Hours and hours trying to figure it out... Copying pieces of bad class to new almost the same class and boom, it works... Never seen anything like it. Anyway, game on:

New command ~~To: <filePath>
All further output is sent to the <filePath> file (recreating the file if needed).
~~To: should be able to be used as many times as you want letting a single script output to any number of files... I'm going to bed